Extract reusable parts of invalid NSEC3 test
